requirements.txt
# Consultoria App Dependencies

# Web framework
streamlit>=1.30.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets integration
gspread>=5.12.0
google-auth>=2.25.0

# HTTP requests
requests>=2.31.0

# Excel file support
openpyxl>=3.1.0

# Text normalization
unidecode>=1.3.0

# Visualization
plotly>=5.18.0

# Progress bars
tqdm>=4.66.0

# Rate limiting (for API clients)
ratelimit>=2.2.0

# Environment variables
python-dotenv>=1.0.0
